Time: 15 ms (83.55%), Space: 10.2 MB (29.00%) - LeetHub

+++ 2114-maximum-number-of-words-found-in-sentences/2114-maximum-number-of-words-found-in-sentences.cpp
@@ -5,7 +5,7 @@ public:
         for(auto str:sentences)
         {
             
-            int wordCnt = (int)count(str.begin(),str.end(),' ') + 1 ; //no leading or trailing spaces
+            int wordCnt = count(str.begin(),str.end(),' ') + 1 ; //no leading or trailing spaces
             
             maxCount = max(maxCount ,wordCnt);
         }
